+++ CHANGELOG.rst
@@ -23,9 +23,11 @@ Unreleased
 * Fixed missing builtin arguments on main ``cms`` management command causing it to crash
 * Fixed template label nested translation
 * Few changes in docs/contributing/code.rst
+* Fixed a bug where the fallback page title whould be returned instead of the one from the current language 
 * Add support for Django 3.2 LTS version
 * removed zh and zh_hans translation (keep zh_CN and zh_TW)
 * Cleaned-up unmaintained translations
+
 
 3.8.0 (2020-10-28)
 ==================